[FEAT] Remove Authorship
